Update the modules folder to fix linter error, and style the input boxes
