+++ src/index.js
@@ -1,7 +1,6 @@
 import './css/style.css';
 import { inputList } from './modules/variableList.js';
-import {displayList} from './modules/showTask.js';
-import addList from './modules/addList.js';
+import { displayList, addList } from './modules/showTask.js';
 
 window.addEventListener('load', () => {
   displayList();
@@ -9,7 +8,6 @@ window.addEventListener('load', () => {
 
 inputList.addEventListener('keypress', (e) => {
   if (e.key === 'Enter') {
-    e.preventDefault;
     addList();
   }
 });
