--- src/modules/editTask.js
@@ -1,9 +1,0 @@
-import { tasksArr, updateList } from "./showTask.js";
-
-const editTask = () => {
-  updateList();
-  const jsonData = JSON.stringify(tasksArr);
-  localStorage.setItem('lists', jsonData);
-}
-
-export default editTask;
